fix(html): split i18n provider from controller instance
