+++ packages/html/src/define/live-video/player.ts
@@ -1,5 +1,5 @@
 import { liveVideoFeatures } from '@videojs/core/dom';
-import { I18nProviderMixin } from '../../i18n/instance';
+import { I18nProviderMixin } from '../../i18n/provider-instance';
 import { MediaContainerElement } from '../../media/container-element';
 import { createPlayer } from '../../player/create-player';
 import { MediaElement } from '../../ui/media-element';
