+++ packages/html/src/i18n/elements.ts
@@ -1,6 +1,7 @@
 import { ReactiveElement } from '@videojs/element';
 
-import { I18nProviderMixin, I18nTextMixin } from './instance';
+import { I18nTextMixin } from './instance';
+import { I18nProviderMixin } from './provider-instance';
 
 export class MediaI18nProviderElement extends I18nProviderMixin(ReactiveElement) {
   static readonly tagName = 'media-i18n-provider';
