+++ packages/html/src/i18n/provider-instance.ts
@@ -1,0 +1,6 @@
+import { createI18nWithBase } from './create-i18n';
+import { i18nBase } from './instance';
+
+const built = createI18nWithBase(i18nBase);
+
+export const I18nProviderMixin = built.ProviderMixin;
